NewFeatures(jumptotoday-description)
Es wird Description Zeile für Tasks hinzugefügt und Button für Jumptotoday Features hinzugefügt.

+++ api/boards.php
@@ -4,23 +4,38 @@ $params = $_GET;
 $payload = file_get_contents('php://input');
 $filename = '../data/' . $params["id"] . '.json';
 
-if($method === "POST") {
-  file_put_contents($filename, $payload);
-}else if($method === "DELETE") {
+if ($method === "POST") {
+  if (!file_exists($filename)) {
+    file_put_contents($filename, $payload);
+    header('Content-Type: application/json');
+    echo $payload;
+  } else {
+    http_response_code(409);
+  }
+
+} else if ($method === "PUT") {
+  if (file_exists($filename)) {
+    file_put_contents($filename, $payload);
+    header('Content-Type: application/json');
+    echo $payload;
+  } else {
+    http_response_code(409);
+  }
+
+} else if ($method === "DELETE") {
   unlink($filename);
-}else if($method === "GET") {
-  if (file_exists($filename))  
-  {  
+
+} else if ($method === "GET") {
+  if (file_exists($filename)) {
     $result = file_get_contents($filename, false);
     header('Content-Type: application/json');
-    echo json_encode($result);
-  } 
-  else 
-  { 
-    file_put_contents($filename, $payload);
-  }  
-}else {
-    
+    echo $result;
+  } else {
+    http_response_code(404);
+  }
+
+} else {
+  http_response_code(405);
 }
 
 ?>
